+++ src/components/sdk/MeetingControls/VideoInputControl.tsx
@@ -7,7 +7,7 @@ import { ControlBarButton } from '../../ui/ControlBar/ControlBarItem';
 import { Camera } from '../../ui/icons';
 import { useMeetingManager } from '../../../providers/MeetingProvider';
 import { useVideoInputs } from '../../../providers/DevicesProvider';
-import { useLocalVideoToggle } from '../../../providers/LocalVideoToggleProvider';
+import { useLocalVideo } from '../../../providers/LocalVideoProvider';
 import { DeviceConfig } from '../../../types';
 import { isOptionActive } from '../../../utils/device-utils';
 import { PopOverItemProps } from '../../ui/PopOver/PopOverItem';
@@ -18,7 +18,7 @@ const VideoInputControl: React.FC = () => {
     additionalDevices: true
   };
   const { devices, selectedDevice } = useVideoInputs(videoInputConfig);
-  const { isVideoEnabled, toggleVideo } = useLocalVideoToggle();
+  const { isVideoEnabled, toggleVideo } = useLocalVideo();
   const dropdownOptions: PopOverItemProps[] = devices.map((device: any) => ({
     children: <span>{device.label}</span>,
     checked: isOptionActive(selectedDevice, device.deviceId),