+++ src/providers/DevicesProvider/VideoInputProvider.tsx
@@ -16,11 +16,10 @@ import { DeviceTypeContext, DeviceConfig } from '../../types';
 import { VIDEO_INPUT } from '../../constants/additional-audio-video-devices';
 import { getFormattedDropdownDeviceOptions } from '../../utils/device-utils';
 
-export type LocalVideoToggleContextType = {
+export type LocalVideoContextType = {
   isVideoEnabled: boolean;
   toggleVideo: () => Promise<void>;
 };
-
 
 const Context = createContext<DeviceTypeContext | null>(null);